Refactor spotify_authorization.dart
This breaks apart SpotifyAuthorization and separates it into different classes to add encapsulation. dotenv_loader.dart, error_handler.dart, and header_builder.dart are created to separate out responsibilities. pubspec.lock and pubspec.yaml are also updated to add a path dependency and the spotify test import statement is updated to adjust for re-arranging the directories.